requirements.txt
requirements-parser==0.1.0
PyGithub